requirements.txt
flask==3.0.3
python-telegram-bot==20.7